corner cases

+++ PA3/server.py
@@ -5,7 +5,7 @@ import time
 # create an INET, STREAMing server socket
 serversocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 # bind the socket to a public host, and a port
-serversocket.bind((socket.gethostname(), 3903))
+serversocket.bind((socket.gethostname(), 3905))
 # become a server socket and queue up to 5 requests
 serversocket.listen(5)
 
@@ -141,34 +141,45 @@ while True:
 
     elif choice == "2":
         while True:
-            print("\nSending index...")
-            clientsocket.send(("\n".join(index.keys()).encode('ascii')))
-            file_choice = clientsocket.recv(100).decode('ascii')
-            if file_choice == '':
+            # if index non-empty
+            if index:
+                print("\nSending index...")
+                log_send = "{};{}".format("\n".join(index.keys()), "1")
+                clientsocket.send((log_send.encode('ascii')))
+                file_choice = clientsocket.recv(100).decode('ascii')
+                if file_choice == '':
+                    print("Client disconnected!")
+                    clientsocket.close()
+                    break
+                # Retrieve from storage
+                retrieve_from_storage(file_choice)
+                # send file size
+                file_size = int(index[file_choice][1])
+                clientsocket.send(str(file_size).encode('ascii'))
+                # wait for server to finish sending log
+                time.sleep(1)
+                print("Sending file %s to client..." % file_choice)
+                f = open("server_tmp/" + file_choice, 'rb')
+
+                while file_size >= 1024:
+                    l = f.read(1024)
+                    clientsocket.send(l)
+                    file_size -= 1024
+
+                if file_size > 0:
+                    l = f.read(file_size)
+                    clientsocket.send(l)
+
+                f.close()
+
+                print("File %s sent to client!" % file_choice)
+                os.remove("server_tmp/" + file_choice)
+                print("Deleted file %s from server." % file_choice)
+
+            else:
+                print("\nNo synced files!")
+                log_send = "{};{}".format("No synced files!", "0")
+                clientsocket.send((log_send.encode('ascii')))
                 print("Client disconnected!")
                 clientsocket.close()
                 break
-            # Retrieve from storage
-            retrieve_from_storage(file_choice)
-            # send file size
-            file_size = int(index[file_choice][1])
-            clientsocket.send(str(file_size).encode('ascii'))
-            # wait for server to finish sending log
-            time.sleep(1)
-            print("Sending file %s to client..." % file_choice)
-            f = open("server_tmp/" + file_choice, 'rb')
-
-            while file_size >= 1024:
-                l = f.read(1024)
-                clientsocket.send(l)
-                file_size -= 1024
-
-            if file_size > 0:
-                l = f.read(file_size)
-                clientsocket.send(l)
-
-            f.close()
-
-            print("File %s sent to client!" % file_choice)
-            os.remove("server_tmp/" + file_choice)
-            print("Deleted file %s from server." % file_choice)
